Add files via upload
The acronym for this program is CWDR (Cakewalk Delay Record.) Please read the PDF for thorough instructions. Other DAWs might be able to use this Desktop program if the same keystrokes for Record, Play and Pause are the same as for Cakewalk. If the DAW has focus, you might need to click twice (depending on how it is used) on the RECORD button within CWDR. I put few notes within the ahk file. I left notes in from code that I borrowed from a submitter on another website being free for any usage. It uses the progress bar characteristics to simulate a Button control that toggles between 2 states when clicked, changing colors and text.